Move create_logger() into intos own module log.py.
