+++ .isort.cfg
@@ -1,2 +1,2 @@
 [settings]
-known_third_party = numpy,pandas
+known_third_party = log,numpy,pandas
