+++ apps/iris/data.py
@@ -1,9 +1,9 @@
 """Module for loading and describing the Iris data set from the web.
 """
 import logging
-import os
 import sys
 
+import log
 import numpy as np
 import pandas as pd
 
@@ -30,41 +30,6 @@ stat_dict = {
 }
 
 _mrsharky_data_path = "https://teaching.mrsharky.com/data/iris.data"
-
-
-def create_logger(
-    log_name: str = __name__,
-    log_level: int = logging.DEBUG,
-    use_file_handler: bool = False,
-) -> logging.Logger:
-    """Configure and return a Python Logger object.
-
-    Args:
-        log_name  (str, optional): Sets new log's name.
-        Defaults to __name__.
-        log_level (int, optional): Sets logging level.
-        Defaults to logging.DEBUG.
-        use_file_handler (bool, optional): Adds a file handler if True.
-        Defaults to False.
-
-    Returns:
-        logging.Logger: A configured Python Logger object.
-
-    Todo:
-        Put this in its own module, eventually in shared library between apps.
-    """
-    logger = logging.getLogger(name=__name__)
-    logger.setLevel(level=log_level)
-    stream_handler = logging.StreamHandler(stream=sys.stdout)
-    logger.addHandler(stream_handler)
-    if use_file_handler is True:
-        # Guarantee logs folder exists: from question 273192 on Stack Overflow.
-        if not os.path.exists("./logs/"):
-            os.makedirs("logs")
-        handler_path = "./logs/" + log_name + ".log"
-        file_handler = logging.FileHandler(handler_path)
-        logger.addHandler(file_handler)
-    return logger
 
 
 def load_data(data_path=_mrsharky_data_path) -> pd.DataFrame:
@@ -119,7 +84,7 @@ def describe_data(data_df: pd.DataFrame) -> pd.DataFrame:
 
 def main() -> int:
     # Setup logger with a helper function
-    logger = create_logger(
+    logger = log.create_logger(
         log_name="iris.data", log_level=logging.DEBUG, use_file_handler=True
     )
     logger.info("Loading Iris data set.")
